+++ z3/src/ast/mod.rs
@@ -18,86 +18,28 @@ use num::{bigint::BigInt, rational::BigRational};
 mod bv;
 mod int;
 mod real;
+mod float;
+mod string;
+mod array;
+mod set;
+mod seq;
+mod dynamic;
+mod datatype;
+mod regexp;
+mod bool;
 
+pub use array::Array;
+pub use bool::Bool;
 pub use bv::BV;
+pub use datatype::Datatype;
+pub use dynamic::Dynamic;
+pub use float::Float;
 pub use int::Int;
 pub use real::Real;
-
-/// [`Ast`] node representing a boolean value.
-pub struct Bool {
-    pub(crate) ctx: Context,
-    pub(crate) z3_ast: Z3_ast,
-}
-
-/// [`Ast`] node representing a float value.
-pub struct Float {
-    pub(crate) ctx: Context,
-    pub(crate) z3_ast: Z3_ast,
-}
-
-/// [`Ast`] node representing a string value.
-pub struct String {
-    pub(crate) ctx: Context,
-    pub(crate) z3_ast: Z3_ast,
-}
-
-/// [`Ast`] node representing an array value.
-/// An array in Z3 is a mapping from indices to values.
-pub struct Array {
-    pub(crate) ctx: Context,
-    pub(crate) z3_ast: Z3_ast,
-}
-
-/// [`Ast`] node representing a set value.
-pub struct Set {
-    pub(crate) ctx: Context,
-    pub(crate) z3_ast: Z3_ast,
-}
-
-/// [`Ast`] node representing a sequence value.
-pub struct Seq {
-    pub(crate) ctx: Context,
-    pub(crate) z3_ast: Z3_ast,
-}
-
-/// [`Ast`] node representing a datatype or enumeration value.
-pub struct Datatype {
-    pub(crate) ctx: Context,
-    pub(crate) z3_ast: Z3_ast,
-}
-
-/// A dynamically typed [`Ast`] node.
-pub struct Dynamic {
-    pub(crate) ctx: Context,
-    pub(crate) z3_ast: Z3_ast,
-}
-
-/// [`Ast`] node representing a regular expression.
-/// ```
-/// use z3::ast;
-/// use z3::{Config, Context, Solver, SatResult};
-///
-/// let cfg = Config::new();
-/// let ctx = &Context::new(&cfg);
-/// let solver = Solver::new(&ctx);
-/// let s = ast::String::new_const(ctx, "s");
-///
-/// // the regexp representing foo[a-c]*
-/// let a = ast::Regexp::concat(ctx, &[
-///     &ast::Regexp::literal(ctx, "foo"),
-///     &ast::Regexp::range(ctx, &'a', &'c').star()
-/// ]);
-/// // the regexp representing [a-z]+
-/// let b = ast::Regexp::range(ctx, &'a', &'z').plus();
-/// // intersection of a and b is non-empty
-/// let intersect = ast::Regexp::intersect(ctx, &[&a, &b]);
-/// solver.assert(&s.regex_matches(&intersect));
-/// assert_eq!(solver.check(), SatResult::Sat);
-/// ```
-pub struct Regexp {
-    pub(crate) ctx: Context,
-    pub(crate) z3_ast: Z3_ast,
-}
+pub use regexp::Regexp;
+pub use seq::Seq;
+pub use set::Set;
+pub use string::String;
 
 macro_rules! unop {
     (
@@ -594,165 +536,10 @@ impl_ast!(Seq);
 impl_from_try_into_dynamic!(Seq, as_seq);
 impl_ast!(Regexp);
 
-impl Float {
-    // Create a 32-bit (IEEE-754) Float [`Ast`] from a rust f32
-    pub fn from_f32(ctx: &Context, value: f32) -> Float {
-        let sort = Sort::float32(ctx);
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_fpa_numeral_float(ctx.z3_ctx.0, value, sort.z3_sort)
-            })
-        }
-    }
-
-    // Create a 364-bit (IEEE-754) Float [`Ast`] from a rust f64
-    pub fn from_f64(ctx: &Context, value: f64) -> Float {
-        let sort = Sort::double(ctx);
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_fpa_numeral_double(ctx.z3_ctx.0, value, sort.z3_sort)
-            })
-        }
-    }
-
-    pub fn as_f64(&self) -> f64 {
-        unsafe { Z3_get_numeral_double(self.ctx.z3_ctx.0, self.z3_ast) }
-    }
-}
-
 impl_ast!(Datatype);
 impl_from_try_into_dynamic!(Datatype, as_datatype);
 
 impl_ast!(Dynamic);
-
-impl Bool {
-    pub fn new_const<S: Into<Symbol>>(ctx: &Context, name: S) -> Bool {
-        let sort = Sort::bool(ctx);
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_const(ctx.z3_ctx.0, name.into().as_z3_symbol(ctx), sort.z3_sort)
-            })
-        }
-    }
-
-    pub fn fresh_const(ctx: &Context, prefix: &str) -> Bool {
-        let sort = Sort::bool(ctx);
-        unsafe {
-            Self::wrap(ctx, {
-                let pp = CString::new(prefix).unwrap();
-                let p = pp.as_ptr();
-                Z3_mk_fresh_const(ctx.z3_ctx.0, p, sort.z3_sort)
-            })
-        }
-    }
-
-    pub fn from_bool(ctx: &Context, b: bool) -> Bool {
-        unsafe {
-            Self::wrap(ctx, {
-                if b {
-                    Z3_mk_true(ctx.z3_ctx.0)
-                } else {
-                    Z3_mk_false(ctx.z3_ctx.0)
-                }
-            })
-        }
-    }
-
-    pub fn as_bool(&self) -> Option<bool> {
-        unsafe {
-            match Z3_get_bool_value(self.ctx.z3_ctx.0, self.z3_ast) {
-                Z3_L_TRUE => Some(true),
-                Z3_L_FALSE => Some(false),
-                _ => None,
-            }
-        }
-    }
-
-    // This doesn't quite fit the trinop! macro because of the generic argty
-    pub fn ite<T>(&self, a: &T, b: &T) -> T
-    where
-        T: Ast,
-    {
-        unsafe {
-            T::wrap(&self.ctx, {
-                Z3_mk_ite(
-                    self.ctx.z3_ctx.0,
-                    self.z3_ast,
-                    a.get_z3_ast(),
-                    b.get_z3_ast(),
-                )
-            })
-        }
-    }
-
-    varop! {
-        and(Z3_mk_and, Self);
-        or(Z3_mk_or, Self);
-    }
-    binop! {
-        xor(Z3_mk_xor, Self);
-        iff(Z3_mk_iff, Self);
-        implies(Z3_mk_implies, Self);
-    }
-    unop! {
-        not(Z3_mk_not, Self);
-    }
-
-    pub fn pb_le(ctx: &Context, values: &[(&Bool, i32)], k: i32) -> Bool {
-        unsafe {
-            Bool::wrap(ctx, {
-                assert!(values.len() <= 0xffffffff);
-                let (values, coefficients): (Vec<Z3_ast>, Vec<i32>) = values
-                    .iter()
-                    .map(|(boolean, coefficient)| (boolean.z3_ast, coefficient))
-                    .unzip();
-                Z3_mk_pble(
-                    ctx.z3_ctx.0,
-                    values.len() as u32,
-                    values.as_ptr(),
-                    coefficients.as_ptr(),
-                    k,
-                )
-            })
-        }
-    }
-    pub fn pb_ge(ctx: &Context, values: &[(&Bool, i32)], k: i32) -> Bool {
-        unsafe {
-            Bool::wrap(ctx, {
-                assert!(values.len() <= 0xffffffff);
-                let (values, coefficients): (Vec<Z3_ast>, Vec<i32>) = values
-                    .iter()
-                    .map(|(boolean, coefficient)| (boolean.z3_ast, coefficient))
-                    .unzip();
-                Z3_mk_pbge(
-                    ctx.z3_ctx.0,
-                    values.len() as u32,
-                    values.as_ptr(),
-                    coefficients.as_ptr(),
-                    k,
-                )
-            })
-        }
-    }
-    pub fn pb_eq(ctx: &Context, values: &[(&Bool, i32)], k: i32) -> Bool {
-        unsafe {
-            Bool::wrap(ctx, {
-                assert!(values.len() <= 0xffffffff);
-                let (values, coefficients): (Vec<Z3_ast>, Vec<i32>) = values
-                    .iter()
-                    .map(|(boolean, coefficient)| (boolean.z3_ast, coefficient))
-                    .unzip();
-                Z3_mk_pbeq(
-                    ctx.z3_ctx.0,
-                    values.len() as u32,
-                    values.as_ptr(),
-                    coefficients.as_ptr(),
-                    k,
-                )
-            })
-        }
-    }
-}
 
 pub fn atmost<'a, I: IntoIterator<Item = &'a Bool>>(ctx: &Context, args: I, k: u32) -> Bool {
     let args: Vec<_> = args.into_iter().map(|f| f.z3_ast).collect();
@@ -789,867 +576,6 @@ fn _atleast(ctx: &Context, args: &[Z3_ast], k: u32) -> Bool {
                 k,
             ),
         )
-    }
-}
-
-impl Float {
-    pub fn new_const<S: Into<Symbol>>(ctx: &Context, name: S, ebits: u32, sbits: u32) -> Float {
-        let sort = Sort::float(ctx, ebits, sbits);
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_const(ctx.z3_ctx.0, name.into().as_z3_symbol(ctx), sort.z3_sort)
-            })
-        }
-    }
-
-    /// Create a 32-bit (IEEE-754) Float [`Ast`].
-    pub fn new_const_float32<S: Into<Symbol>>(ctx: &Context, name: S) -> Float {
-        let sort = Sort::float32(ctx);
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_const(ctx.z3_ctx.0, name.into().as_z3_symbol(ctx), sort.z3_sort)
-            })
-        }
-    }
-
-    /// Create a 64-bit (IEEE-754) Float [`Ast`].
-    pub fn new_const_double<S: Into<Symbol>>(ctx: &Context, name: S) -> Float {
-        let sort = Sort::double(ctx);
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_const(ctx.z3_ctx.0, name.into().as_z3_symbol(ctx), sort.z3_sort)
-            })
-        }
-    }
-
-    pub fn fresh_const(ctx: &Context, prefix: &str, ebits: u32, sbits: u32) -> Float {
-        let sort = Sort::float(ctx, ebits, sbits);
-        unsafe {
-            Self::wrap(ctx, {
-                let pp = CString::new(prefix).unwrap();
-                let p = pp.as_ptr();
-                Z3_mk_fresh_const(ctx.z3_ctx.0, p, sort.z3_sort)
-            })
-        }
-    }
-
-    pub fn fresh_const_float32(ctx: &Context, prefix: &str) -> Float {
-        let sort = Sort::float32(ctx);
-        unsafe {
-            Self::wrap(ctx, {
-                let pp = CString::new(prefix).unwrap();
-                let p = pp.as_ptr();
-                Z3_mk_fresh_const(ctx.z3_ctx.0, p, sort.z3_sort)
-            })
-        }
-    }
-
-    pub fn fresh_const_double(ctx: &Context, prefix: &str) -> Float {
-        let sort = Sort::double(ctx);
-        unsafe {
-            Self::wrap(ctx, {
-                let pp = CString::new(prefix).unwrap();
-                let p = pp.as_ptr();
-                Z3_mk_fresh_const(ctx.z3_ctx.0, p, sort.z3_sort)
-            })
-        }
-    }
-
-    // returns RoundingMode towards zero
-    pub fn round_towards_zero(ctx: &Context) -> Float {
-        unsafe { Self::wrap(ctx, Z3_mk_fpa_round_toward_zero(ctx.z3_ctx.0)) }
-    }
-
-    // returns RoundingMode towards negative
-    pub fn round_towards_negative(ctx: &Context) -> Float {
-        unsafe { Self::wrap(ctx, Z3_mk_fpa_round_toward_negative(ctx.z3_ctx.0)) }
-    }
-
-    // returns RoundingMode towards positive
-    pub fn round_towards_positive(ctx: &Context) -> Float {
-        unsafe { Self::wrap(ctx, Z3_mk_fpa_round_toward_positive(ctx.z3_ctx.0)) }
-    }
-
-    // Add two floats of the same size, rounding towards zero
-    pub fn add_towards_zero(&self, other: &Self) -> Float {
-        Self::round_towards_zero(&self.ctx).add(self, other)
-    }
-
-    // Subtract two floats of the same size, rounding towards zero
-    pub fn sub_towards_zero(&self, other: &Self) -> Float {
-        Self::round_towards_zero(&self.ctx).sub(self, other)
-    }
-
-    // Multiply two floats of the same size, rounding towards zero
-    pub fn mul_towards_zero(&self, other: &Self) -> Float {
-        Self::round_towards_zero(&self.ctx).mul(self, other)
-    }
-
-    // Divide two floats of the same size, rounding towards zero
-    pub fn div_towards_zero(&self, other: &Self) -> Float {
-        Self::round_towards_zero(&self.ctx).div(self, other)
-    }
-
-    // Convert to IEEE-754 bit-vector
-    pub fn to_ieee_bv(&self) -> BV {
-        unsafe {
-            BV::wrap(
-                &self.ctx,
-                Z3_mk_fpa_to_ieee_bv(self.ctx.z3_ctx.0, self.z3_ast),
-            )
-        }
-    }
-
-    unop! {
-        unary_abs(Z3_mk_fpa_abs, Self);
-        unary_neg(Z3_mk_fpa_neg, Self);
-        is_infinite(Z3_mk_fpa_is_infinite, Bool);
-        is_normal(Z3_mk_fpa_is_normal, Bool);
-        is_subnormal(Z3_mk_fpa_is_subnormal, Bool);
-        is_zero(Z3_mk_fpa_is_zero, Bool);
-        is_nan(Z3_mk_fpa_is_nan, Bool);
-    }
-    binop! {
-        lt(Z3_mk_fpa_lt, Bool);
-        le(Z3_mk_fpa_leq, Bool);
-        gt(Z3_mk_fpa_gt, Bool);
-        ge(Z3_mk_fpa_geq, Bool);
-    }
-    trinop! {
-        add(Z3_mk_fpa_add, Self);
-        sub(Z3_mk_fpa_sub, Self);
-        mul(Z3_mk_fpa_mul, Self);
-        div(Z3_mk_fpa_div, Self);
-    }
-}
-
-impl String {
-    /// Creates a new constant using the built-in string sort
-    pub fn new_const<S: Into<Symbol>>(ctx: &Context, name: S) -> String {
-        let sort = Sort::string(ctx);
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_const(ctx.z3_ctx.0, name.into().as_z3_symbol(ctx), sort.z3_sort)
-            })
-        }
-    }
-
-    /// Creates a fresh constant using the built-in string sort
-    pub fn fresh_const(ctx: &Context, prefix: &str) -> String {
-        let sort = Sort::string(ctx);
-        unsafe {
-            Self::wrap(ctx, {
-                let pp = CString::new(prefix).unwrap();
-                let p = pp.as_ptr();
-                Z3_mk_fresh_const(ctx.z3_ctx.0, p, sort.z3_sort)
-            })
-        }
-    }
-
-    /// Creates a Z3 constant string from a `&str`
-    pub fn from_str(ctx: &Context, string: &str) -> Result<String, std::ffi::NulError> {
-        let string = CString::new(string)?;
-        Ok(unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_string(ctx.z3_ctx.0, string.as_c_str().as_ptr())
-            })
-        })
-    }
-
-    /// Retrieves the underlying `std::string::String`
-    ///
-    /// If this is not a constant `z3::ast::String`, return `None`.
-    ///
-    /// Note that `to_string()` provided by `std::string::ToString` (which uses
-    /// `std::fmt::Display`) returns an escaped string. In contrast,
-    /// `z3::ast::String::from_str(&ctx, s).unwrap().as_string()` returns a
-    /// `String` equal to the original value.
-    pub fn as_string(&self) -> Option<std::string::String> {
-        let z3_ctx = self.get_ctx().z3_ctx.0;
-        unsafe {
-            let bytes = Z3_get_string(z3_ctx, self.get_z3_ast());
-            if bytes.is_null() {
-                None
-            } else {
-                Some(CStr::from_ptr(bytes).to_string_lossy().into_owned())
-            }
-        }
-    }
-
-    /// Retrieve the substring of length 1 positioned at `index`.
-    ///
-    /// # Examples
-    /// ```
-    /// # use z3::{Config, Context, Solver};
-    /// # use z3::ast::{Ast as _, Int};
-    /// #
-    /// # let cfg = Config::new();
-    /// # let ctx = Context::new(&cfg);
-    /// # let solver = Solver::new(&ctx);
-    /// #
-    /// let s = z3::ast::String::fresh_const(&ctx, "");
-    ///
-    /// solver.assert(
-    ///     &s.at(&Int::from_u64(&ctx, 0))
-    ///         ._eq(&z3::ast::String::from_str(&ctx, "a").unwrap())
-    /// );
-    /// assert_eq!(solver.check(), z3::SatResult::Sat);
-    /// ```
-    pub fn at(&self, index: &Int) -> Self {
-        unsafe {
-            Self::wrap(
-                &self.ctx,
-                Z3_mk_seq_at(self.ctx.z3_ctx.0, self.z3_ast, index.z3_ast),
-            )
-        }
-    }
-
-    /// Retrieve the substring of length `length` starting at `offset`.
-    ///
-    /// # Examples
-    /// ```
-    /// # use z3::{Config, Context, Solver, SatResult};
-    /// # use z3::ast::{Ast as _, Int, String};
-    /// #
-    /// # let cfg = Config::new();
-    /// # let ctx = Context::new(&cfg);
-    /// # let solver = Solver::new(&ctx);
-    /// #
-    /// let s = String::from_str(&ctx, "abc").unwrap();
-    /// let sub = String::fresh_const(&ctx, "");
-    ///
-    /// solver.assert(
-    ///     &sub._eq(
-    ///         &s.substr(
-    ///             &Int::from_u64(&ctx, 1),
-    ///             &Int::from_u64(&ctx, 2),
-    ///         )
-    ///     )
-    /// );
-    ///
-    /// assert_eq!(solver.check(), SatResult::Sat);
-    /// assert_eq!(
-    ///     solver
-    ///         .get_model()
-    ///         .unwrap()
-    ///         .eval(&sub, true)
-    ///         .unwrap()
-    ///         .as_string()
-    ///         .unwrap()
-    ///         .as_str(),
-    ///     "bc",
-    /// );
-    /// ```
-    pub fn substr(&self, offset: &Int, length: &Int) -> Self {
-        unsafe {
-            Self::wrap(
-                &self.ctx,
-                Z3_mk_seq_extract(self.ctx.z3_ctx.0, self.z3_ast, offset.z3_ast, length.z3_ast),
-            )
-        }
-    }
-
-    /// Checks if this string matches a `z3::ast::Regexp`
-    pub fn regex_matches(&self, regex: &Regexp) -> Bool {
-        assert!(self.ctx == regex.ctx);
-        unsafe {
-            Bool::wrap(
-                &self.ctx,
-                Z3_mk_seq_in_re(self.ctx.z3_ctx.0, self.get_z3_ast(), regex.get_z3_ast()),
-            )
-        }
-    }
-
-    varop! {
-        /// Appends the argument strings to `Self`
-        concat(Z3_mk_seq_concat, String);
-    }
-
-    unop! {
-        /// Gets the length of `Self`.
-        length(Z3_mk_seq_length, Int);
-    }
-
-    binop! {
-        /// Checks whether `Self` contains a substring
-        contains(Z3_mk_seq_contains, Bool);
-        /// Checks whether `Self` is a prefix of the argument
-        prefix(Z3_mk_seq_prefix, Bool);
-        /// Checks whether `Self` is a suffix of the argument
-        suffix(Z3_mk_seq_suffix, Bool);
-    }
-}
-
-impl Array {
-    /// Create an `Array` which maps from indices of the `domain` `Sort` to
-    /// values of the `range` `Sort`.
-    ///
-    /// All values in the `Array` will be unconstrained.
-    pub fn new_const<S: Into<Symbol>>(
-        ctx: &Context,
-        name: S,
-        domain: &Sort,
-        range: &Sort,
-    ) -> Array {
-        let sort = Sort::array(ctx, domain, range);
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_const(ctx.z3_ctx.0, name.into().as_z3_symbol(ctx), sort.z3_sort)
-            })
-        }
-    }
-
-    pub fn fresh_const(ctx: &Context, prefix: &str, domain: &Sort, range: &Sort) -> Array {
-        let sort = Sort::array(ctx, domain, range);
-        unsafe {
-            Self::wrap(ctx, {
-                let pp = CString::new(prefix).unwrap();
-                let p = pp.as_ptr();
-                Z3_mk_fresh_const(ctx.z3_ctx.0, p, sort.z3_sort)
-            })
-        }
-    }
-
-    /// Create a "constant array", that is, an `Array` initialized so that all of the
-    /// indices in the `domain` map to the given value `val`
-    pub fn const_array<A>(ctx: &Context, domain: &Sort, val: &A) -> Array
-    where
-        A: Ast,
-    {
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_const_array(ctx.z3_ctx.0, domain.z3_sort, val.get_z3_ast())
-            })
-        }
-    }
-
-    /// Get the value at a given index in the array.
-    ///
-    /// Note that the `index` _must be_ of the array's `domain` sort.
-    /// The return type will be of the array's `range` sort.
-    //
-    // We avoid the binop! macro because the argument has a non-Self type
-    pub fn select<A>(&self, index: &A) -> Dynamic
-    where
-        A: Ast,
-    {
-        // TODO: We could validate here that the index is of the correct type.
-        // This would require us either to keep around the original `domain` argument
-        // from when the Array was constructed, or to do an additional Z3 query
-        // to find the domain sort first.
-        // But if we did this check ourselves, we'd just panic, so it doesn't seem
-        // like a huge advantage over just letting Z3 panic itself when it discovers the
-        // problem.
-        // This way we also avoid the redundant check every time this method is called.
-        unsafe {
-            Dynamic::wrap(&self.ctx, {
-                Z3_mk_select(self.ctx.z3_ctx.0, self.z3_ast, index.get_z3_ast())
-            })
-        }
-    }
-
-    /// n-ary Array read. `idxs` are the indices of the array that gets read.
-    /// This is useful for applying lambdas.
-    pub fn select_n(&self, idxs: &[&dyn Ast]) -> Dynamic {
-        let idxs: Vec<_> = idxs.iter().map(|idx| idx.get_z3_ast()).collect();
-
-        unsafe {
-            Dynamic::wrap(&self.ctx, {
-                Z3_mk_select_n(
-                    self.ctx.z3_ctx.0,
-                    self.z3_ast,
-                    idxs.len().try_into().unwrap(),
-                    idxs.as_ptr() as *const Z3_ast,
-                )
-            })
-        }
-    }
-
-    /// Update the value at a given index in the array.
-    ///
-    /// Note that the `index` _must be_ of the array's `domain` sort,
-    /// and the `value` _must be_ of the array's `range` sort.
-    //
-    // We avoid the trinop! macro because the arguments have non-Self types
-    pub fn store<A1, A2>(&self, index: &A1, value: &A2) -> Self
-    where
-        A1: Ast,
-        A2: Ast,
-    {
-        unsafe {
-            Self::wrap(&self.ctx, {
-                Z3_mk_store(
-                    self.ctx.z3_ctx.0,
-                    self.z3_ast,
-                    index.get_z3_ast(),
-                    value.get_z3_ast(),
-                )
-            })
-        }
-    }
-
-    /// Returns true if the array is a const array (i.e. `a.is_const_array() => exists v, forall i. select(a, i) == v`)
-    ///
-    /// # Examples
-    /// ```
-    /// # use z3::{ast, Config, Context, ast::{Array, Int}, Sort};
-    /// # use z3::ast::Ast;
-    /// # use std::convert::TryInto;
-    /// # let cfg = Config::new();
-    /// # let ctx = Context::new(&cfg);
-    /// let arr = Array::const_array(&ctx, &Sort::int(&ctx), &Int::from_u64(&ctx, 9));
-    /// assert!(arr.is_const_array());
-    /// let arr2 = Array::fresh_const(&ctx, "a", &Sort::int(&ctx), &Sort::int(&ctx));
-    /// assert!(!arr2.is_const_array());
-    /// ```
-    pub fn is_const_array(&self) -> bool {
-        // python:
-        // is_app_of(a, Z3_OP_CONST_ARRAY)
-        // >> is_app(a) and a.decl().kind() == Z3_OP_CONST_ARRAY
-        self.is_app() && matches!(self.decl().kind(), DeclKind::CONST_ARRAY)
-    }
-}
-
-impl Set {
-    pub fn new_const<S: Into<Symbol>>(ctx: &Context, name: S, eltype: &Sort) -> Set {
-        let sort = Sort::set(ctx, eltype);
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_const(ctx.z3_ctx.0, name.into().as_z3_symbol(ctx), sort.z3_sort)
-            })
-        }
-    }
-
-    pub fn fresh_const(ctx: &Context, prefix: &str, eltype: &Sort) -> Set {
-        let sort = Sort::set(ctx, eltype);
-        unsafe {
-            Self::wrap(ctx, {
-                let pp = CString::new(prefix).unwrap();
-                let p = pp.as_ptr();
-                Z3_mk_fresh_const(ctx.z3_ctx.0, p, sort.z3_sort)
-            })
-        }
-    }
-
-    /// Creates a set that maps the domain to false by default
-    pub fn empty(ctx: &Context, domain: &Sort) -> Set {
-        unsafe { Self::wrap(ctx, Z3_mk_empty_set(ctx.z3_ctx.0, domain.z3_sort)) }
-    }
-
-    /// Add an element to the set.
-    ///
-    /// Note that the `element` _must be_ of the `Set`'s `eltype` sort.
-    //
-    // We avoid the binop! macro because the argument has a non-Self type
-    pub fn add<A>(&self, element: &A) -> Set
-    where
-        A: Ast,
-    {
-        unsafe {
-            Self::wrap(&self.ctx, {
-                Z3_mk_set_add(self.ctx.z3_ctx.0, self.z3_ast, element.get_z3_ast())
-            })
-        }
-    }
-
-    /// Remove an element from the set.
-    ///
-    /// Note that the `element` _must be_ of the `Set`'s `eltype` sort.
-    //
-    // We avoid the binop! macro because the argument has a non-Self type
-    pub fn del<A>(&self, element: &A) -> Set
-    where
-        A: Ast,
-    {
-        unsafe {
-            Self::wrap(&self.ctx, {
-                Z3_mk_set_del(self.ctx.z3_ctx.0, self.z3_ast, element.get_z3_ast())
-            })
-        }
-    }
-
-    /// Check if an item is a member of the set.
-    ///
-    /// Note that the `element` _must be_ of the `Set`'s `eltype` sort.
-    //
-    // We avoid the binop! macro because the argument has a non-Self type
-    pub fn member<A>(&self, element: &A) -> Bool
-    where
-        A: Ast,
-    {
-        unsafe {
-            Bool::wrap(&self.ctx, {
-                Z3_mk_set_member(self.ctx.z3_ctx.0, element.get_z3_ast(), self.z3_ast)
-            })
-        }
-    }
-
-    varop! {
-        /// Take the intersection of a list of sets.
-        intersect(Z3_mk_set_intersect, Self);
-        /// Take the union of a list of sets.
-        set_union(Z3_mk_set_union, Self);
-    }
-    unop! {
-        /// Take the complement of the set.
-        complement(Z3_mk_set_complement, Self);
-    }
-    binop! {
-        /// Check if the set is a subset of another set.
-        set_subset(Z3_mk_set_subset, Bool);
-        /// Take the set difference between two sets.
-        difference(Z3_mk_set_difference, Self);
-    }
-}
-
-impl Seq {
-    pub fn new_const<S: Into<Symbol>>(ctx: &Context, name: S, eltype: &Sort) -> Self {
-        let sort = Sort::seq(ctx, eltype);
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_const(ctx.z3_ctx.0, name.into().as_z3_symbol(ctx), sort.z3_sort)
-            })
-        }
-    }
-
-    pub fn fresh_const(ctx: &Context, prefix: &str, eltype: &Sort) -> Self {
-        let sort = Sort::seq(ctx, eltype);
-        unsafe {
-            Self::wrap(ctx, {
-                let pp = CString::new(prefix).unwrap();
-                let p = pp.as_ptr();
-                Z3_mk_fresh_const(ctx.z3_ctx.0, p, sort.z3_sort)
-            })
-        }
-    }
-
-    /// Create a unit sequence of `a`.
-    pub fn unit<A: Ast>(ctx: &Context, a: &A) -> Self {
-        unsafe { Self::wrap(ctx, Z3_mk_seq_unit(ctx.z3_ctx.0, a.get_z3_ast())) }
-    }
-
-    /// Retrieve the unit sequence positioned at position `index`.
-    /// Use [`Seq::nth`] to get just the element.
-    pub fn at(&self, index: &Int) -> Self {
-        unsafe {
-            Self::wrap(
-                &self.ctx,
-                Z3_mk_seq_at(self.ctx.z3_ctx.0, self.z3_ast, index.z3_ast),
-            )
-        }
-    }
-
-    /// Retrieve the element positioned at position `index`.
-    ///
-    /// # Examples
-    /// ```
-    /// # use z3::{ast, Config, Context, Solver, Sort};
-    /// # use z3::ast::{Ast, Bool, Int, Seq};
-    /// # let cfg = Config::new();
-    /// # let ctx = Context::new(&cfg);
-    /// # let solver = Solver::new(&ctx);
-    /// let seq = Seq::fresh_const(&ctx, "", &Sort::bool(&ctx));
-    ///
-    /// solver.assert(
-    ///     &seq.nth(&Int::from_u64(&ctx, 0))
-    ///         .simplify()
-    ///         .as_bool()
-    ///         .unwrap()
-    ///         ._eq(&Bool::from_bool(&ctx, true))
-    /// );
-    /// ```
-    pub fn nth(&self, index: &Int) -> Dynamic {
-        unsafe {
-            Dynamic::wrap(
-                &self.ctx,
-                Z3_mk_seq_nth(self.ctx.z3_ctx.0, self.z3_ast, index.z3_ast),
-            )
-        }
-    }
-
-    pub fn length(&self) -> Int {
-        unsafe { Int::wrap(&self.ctx, Z3_mk_seq_length(self.ctx.z3_ctx.0, self.z3_ast)) }
-    }
-
-    varop! {
-        /// Concatenate sequences.
-        concat(Z3_mk_seq_concat, Self);
-    }
-}
-
-impl Dynamic {
-    pub fn from_ast(ast: &dyn Ast) -> Self {
-        unsafe { Self::wrap(ast.get_ctx(), ast.get_z3_ast()) }
-    }
-
-    pub fn new_const<S: Into<Symbol>>(ctx: &Context, name: S, sort: &Sort) -> Self {
-        unsafe {
-            Self::wrap(
-                ctx,
-                Z3_mk_const(ctx.z3_ctx.0, name.into().as_z3_symbol(ctx), sort.z3_sort),
-            )
-        }
-    }
-
-    pub fn fresh_const(ctx: &Context, prefix: &str, sort: &Sort) -> Self {
-        unsafe {
-            Self::wrap(ctx, {
-                let pp = CString::new(prefix).unwrap();
-                let p = pp.as_ptr();
-                Z3_mk_fresh_const(ctx.z3_ctx.0, p, sort.z3_sort)
-            })
-        }
-    }
-
-    pub fn sort_kind(&self) -> SortKind {
-        unsafe {
-            Z3_get_sort_kind(
-                self.ctx.z3_ctx.0,
-                Z3_get_sort(self.ctx.z3_ctx.0, self.z3_ast),
-            )
-        }
-    }
-
-    /// Returns `None` if the `Dynamic` is not actually a `Bool`
-    pub fn as_bool(&self) -> Option<Bool> {
-        match self.sort_kind() {
-            SortKind::Bool => Some(unsafe { Bool::wrap(&self.ctx, self.z3_ast) }),
-            _ => None,
-        }
-    }
-
-    /// Returns `None` if the `Dynamic` is not actually an `Int`
-    pub fn as_int(&self) -> Option<Int> {
-        match self.sort_kind() {
-            SortKind::Int => Some(unsafe { Int::wrap(&self.ctx, self.z3_ast) }),
-            _ => None,
-        }
-    }
-
-    /// Returns `None` if the `Dynamic` is not actually a `Real`
-    pub fn as_real(&self) -> Option<Real> {
-        match self.sort_kind() {
-            SortKind::Real => Some(unsafe { Real::wrap(&self.ctx, self.z3_ast) }),
-            _ => None,
-        }
-    }
-
-    /// Returns `None` if the `Dynamic` is not actually a `Float`
-    pub fn as_float(&self) -> Option<Float> {
-        match self.sort_kind() {
-            SortKind::FloatingPoint => Some(unsafe { Float::wrap(&self.ctx, self.z3_ast) }),
-            _ => None,
-        }
-    }
-
-    /// Returns `None` if the `Dynamic` is not actually a `String`
-    pub fn as_string(&self) -> Option<String> {
-        unsafe {
-            if Z3_is_string_sort(
-                self.ctx.z3_ctx.0,
-                Z3_get_sort(self.ctx.z3_ctx.0, self.z3_ast),
-            ) {
-                Some(String::wrap(&self.ctx, self.z3_ast))
-            } else {
-                None
-            }
-        }
-    }
-
-    /// Returns `None` if the `Dynamic` is not actually a `BV`
-    pub fn as_bv(&self) -> Option<BV> {
-        match self.sort_kind() {
-            SortKind::BV => Some(unsafe { BV::wrap(&self.ctx, self.z3_ast) }),
-            _ => None,
-        }
-    }
-
-    /// Returns `None` if the `Dynamic` is not actually an `Array`
-    pub fn as_array(&self) -> Option<Array> {
-        match self.sort_kind() {
-            SortKind::Array => Some(unsafe { Array::wrap(&self.ctx, self.z3_ast) }),
-            _ => None,
-        }
-    }
-
-    /// Returns `None` if the `Dynamic` is not actually a `Set`
-    pub fn as_set(&self) -> Option<Set> {
-        unsafe {
-            match self.sort_kind() {
-                SortKind::Array => {
-                    match Z3_get_sort_kind(
-                        self.ctx.z3_ctx.0,
-                        Z3_get_array_sort_range(
-                            self.ctx.z3_ctx.0,
-                            Z3_get_sort(self.ctx.z3_ctx.0, self.z3_ast),
-                        ),
-                    ) {
-                        SortKind::Bool => Some(Set::wrap(&self.ctx, self.z3_ast)),
-                        _ => None,
-                    }
-                }
-                _ => None,
-            }
-        }
-    }
-
-    /// Returns `None` if the `Dynamic` is not actually a `Seq`.
-    pub fn as_seq(&self) -> Option<Seq> {
-        match self.sort_kind() {
-            SortKind::Seq => Some(unsafe { Seq::wrap(&self.ctx, self.z3_ast) }),
-            _ => None,
-        }
-    }
-
-    /// Returns `None` if the `Dynamic` is not actually a `Datatype`
-    pub fn as_datatype(&self) -> Option<Datatype> {
-        match self.sort_kind() {
-            SortKind::Datatype => Some(unsafe { Datatype::wrap(&self.ctx, self.z3_ast) }),
-            _ => None,
-        }
-    }
-}
-
-impl Datatype {
-    pub fn new_const<S: Into<Symbol>>(ctx: &Context, name: S, sort: &Sort) -> Self {
-        assert_eq!(ctx, &sort.ctx);
-        assert_eq!(sort.kind(), SortKind::Datatype);
-
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_const(ctx.z3_ctx.0, name.into().as_z3_symbol(ctx), sort.z3_sort)
-            })
-        }
-    }
-
-    pub fn fresh_const(ctx: &Context, prefix: &str, sort: &Sort) -> Self {
-        assert_eq!(ctx, &sort.ctx);
-        assert_eq!(sort.kind(), SortKind::Datatype);
-
-        unsafe {
-            Self::wrap(ctx, {
-                let pp = CString::new(prefix).unwrap();
-                let p = pp.as_ptr();
-                Z3_mk_fresh_const(ctx.z3_ctx.0, p, sort.z3_sort)
-            })
-        }
-    }
-}
-
-impl Regexp {
-    /// Creates a regular expression that recognizes the string given as parameter
-    pub fn literal(ctx: &Context, s: &str) -> Self {
-        unsafe {
-            Self::wrap(ctx, {
-                let c_str = CString::new(s).unwrap();
-                Z3_mk_seq_to_re(ctx.z3_ctx.0, Z3_mk_string(ctx.z3_ctx.0, c_str.as_ptr()))
-            })
-        }
-    }
-
-    /// Creates a regular expression that recognizes a character in the specified range (e.g.
-    /// `[a-z]`)
-    pub fn range(ctx: &Context, lo: &char, hi: &char) -> Self {
-        unsafe {
-            Self::wrap(ctx, {
-                let lo_cs = CString::new(lo.to_string()).unwrap();
-                let hi_cs = CString::new(hi.to_string()).unwrap();
-                let lo_z3s = Z3_mk_string(ctx.z3_ctx.0, lo_cs.as_ptr());
-                Z3_inc_ref(ctx.z3_ctx.0, lo_z3s);
-                let hi_z3s = Z3_mk_string(ctx.z3_ctx.0, hi_cs.as_ptr());
-                Z3_inc_ref(ctx.z3_ctx.0, hi_z3s);
-
-                let ret = Z3_mk_re_range(ctx.z3_ctx.0, lo_z3s, hi_z3s);
-                Z3_dec_ref(ctx.z3_ctx.0, lo_z3s);
-                Z3_dec_ref(ctx.z3_ctx.0, hi_z3s);
-                ret
-            })
-        }
-    }
-
-    /// Creates a regular expression that recognizes this regular expression `lo` to `hi` times (e.g. `a{2,3}`)
-    pub fn r#loop(&self, lo: u32, hi: u32) -> Self {
-        unsafe {
-            Self::wrap(&self.ctx, {
-                Z3_mk_re_loop(self.ctx.z3_ctx.0, self.z3_ast, lo, hi)
-            })
-        }
-    }
-
-    /// Creates a regular expression that recognizes this regular expression
-    /// n number of times
-    /// Requires Z3 4.8.15 or later.
-    pub fn power(&self, n: u32) -> Self {
-        unsafe {
-            Self::wrap(&self.ctx, {
-                Z3_mk_re_power(self.ctx.z3_ctx.0, self.z3_ast, n)
-            })
-        }
-    }
-
-    /// Creates a regular expression that recognizes all sequences
-    pub fn full(ctx: &Context) -> Self {
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_re_full(
-                    ctx.z3_ctx.0,
-                    Z3_mk_re_sort(ctx.z3_ctx.0, Z3_mk_string_sort(ctx.z3_ctx.0)),
-                )
-            })
-        }
-    }
-
-    /// Creates a regular expression that accepts all singleton sequences of the characters
-    /// Requires Z3 4.8.13 or later.
-    pub fn allchar(ctx: &Context) -> Self {
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_re_allchar(
-                    ctx.z3_ctx.0,
-                    Z3_mk_re_sort(ctx.z3_ctx.0, Z3_mk_string_sort(ctx.z3_ctx.0)),
-                )
-            })
-        }
-    }
-
-    /// Creates a regular expression that doesn't recognize any sequences
-    pub fn empty(ctx: &Context) -> Self {
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_re_empty(
-                    ctx.z3_ctx.0,
-                    Z3_mk_re_sort(ctx.z3_ctx.0, Z3_mk_string_sort(ctx.z3_ctx.0)),
-                )
-            })
-        }
-    }
-
-    unop! {
-       /// Creates a regular expression that recognizes this regular expression one or more times (e.g. `a+`)
-       plus(Z3_mk_re_plus, Self);
-       /// Creates a regular expression that recognizes this regular expression any number of times
-       /// (Kleene star, e.g. `a*`)
-       star(Z3_mk_re_star, Self);
-       /// Creates a regular expression that recognizes any sequence that this regular expression
-       /// doesn't
-       complement(Z3_mk_re_complement, Self);
-       /// Creates a regular expression that optionally accepts this regular expression (e.g. `a?`)
-       option(Z3_mk_re_option, Self);
-    }
-    binop! {
-        /// Creates a difference regular expression
-        /// Requires Z3 4.8.14 or later.
-        diff(Z3_mk_re_diff, Self);
-    }
-    varop! {
-       /// Concatenates regular expressions
-        concat(Z3_mk_re_concat, Self);
-       /// Creates a regular expression that recognizes sequences that any of the regular
-       /// expressions given as parameters recognize
-        union(Z3_mk_re_union, Self);
-        /// Creates a regular expression that only recognizes sequences that all of the parameters
-        /// recognize
-        intersect(Z3_mk_re_intersect, Self);
     }
 }
 
@@ -1874,7 +800,7 @@ pub fn quantifier_const(
 /// # Examples
 /// ```
 /// # use z3::{
-/// #     ast::{lambda_const, Ast as _, Int, Dynamic},
+/// #     ast::{lambda_const, Ast as _, Int },
 /// #     Config, Context, Solver, SatResult,
 /// # };
 /// #
