+++ z3/tests/ops.rs
@@ -1,7 +1,12 @@
 use z3::{
-    Config, Context, DeclKind, FuncDecl, SatResult, Solver, Sort, ast,
-    ast::{Array, Ast, AstKind, BV, Bool, Dynamic, Float, Int, Real},
+    ast, ast::{Ast, AstKind, Int, Real, BV}, Config, Context, DeclKind, FuncDecl, SatResult, Solver,
+    Sort,
 };
+use z3::ast::Array;
+use z3::ast::Bool;
+use z3::ast::Dynamic;
+use z3::ast::Float;
+use z3::ast::Set;
 
 #[test]
 fn test_bv_ops() {
@@ -347,7 +352,7 @@ fn test_ast_attributes() {
     assert_ast_attributes(&ast::String::new_const(&ctx, "st"), true);
 
     let int_expr = Int::new_const(&ctx, "i");
-    let set_expr = ast::Set::new_const(&ctx, "set", &Sort::int(&ctx));
+    let set_expr = Set::new_const(&ctx, "set", &Sort::int(&ctx));
     assert_ast_attributes(&int_expr, true);
     assert_ast_attributes(&set_expr, true);
     assert_ast_attributes(&set_expr.add(&Dynamic::from_ast(&int_expr)), false);
